Impl GeomEventProcessor for Bbox
